6-7-18 10:57

+++ hw5/VM.cpp
@@ -11,8 +11,8 @@ VM::VM() : sram(SRAM, HOLD) {
 	nextVictimPage = 0;
 }
 
-float VM::getFaultRate() {
-	return (float) numberofPageFaults / numberofMemoryReferences;
+double VM::getFaultRate() {
+	return (double) numberofPageFaults / (double) numberofMemoryReferences;
 }
 
 void VM::resetFaultRate() {
